% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABO}
\alias{ABO}
\title{ABO blood type allelic frequencies}
\format{A data frame with 3 rows and 2 columns:
\itemize{
  \item abo: AB0 alleles
  \item freq: allelic relative frequencies
  }}
\usage{
ABO
}
\description{
Data frame with ABO allelic frequencies from Portuguese blood donors
}
\examples{
\dontrun{
 ABO
}
}
\references{
Duran J, Chabert T, Rodrigues F, Pestana D.
Distribuição dos grupos sanguíneos na população portuguesa. AB0 29, 2007: 5-17
(\href{http://www.ipst.pt/files/IPST/INFORMACAO_DOCUMENTACAO/AB0_29_2007_pag_5a17.pdf}{IPST})
}
\keyword{datasets}
