% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acHLA}
\alias{acHLA}
\title{HLA antibodies}
\format{A data frame with 2 columns:
\itemize{
 \item ID: candidates identification
 \item acs: antibodies anti-HLA
}}
\usage{
acHLA
}
\description{
Data frame with HLA antibodies from wich patients are sensitized
}
\examples{
\dontrun{
 acHLA
}
}
\references{
Bruno A Lima, Helena Alves. Assessment of a Portuguese Panel Reactive Antibody calculator. HLA 91(5), 2018: pp374
 (\href{http://bioestatisticas.wixsite.com/bioestatisticas/publicacoes}{OB})
}
\keyword{datasets}
