% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpra.R
\name{cpra1}
\alias{cpra1}
\title{cPRA value for a single candidate}
\usage{
cpra1(id = 1, sensib = acHLA, A_freq = HLA_A, B_freq = HLA_B,
  DR_freq = HLA_DR, AB_freq = HLA_AB, ADR_freq = HLA_ADR,
  BDR_freq = HLA_BDR, ABDR_freq = HLA_ABDR)
}
\arguments{
\item{id}{a numeric valic with patients identification}

\item{sensib}{a 2 columns data frame with patients identification 'ID'
and HLA antibodies 'acs'}

\item{A_freq}{a 2 columns dataframe with HLA-A alelic frequencies.
A column (named 'AA') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{B_freq}{a 2 columns dataframe with HLA-B alelic frequencies.
A column (named 'BB') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{DR_freq}{a 2 columns dataframe with HLA-DR alelic frequencies.
A column (named 'DDR') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{AB_freq}{a 3 columns dataframe with HLA-A-B haplotypic frequencies.
A column (named 'AA') with HLA-A alleles, another (named 'BB') with HLA-B alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{ADR_freq}{a 3 columns dataframe with HLA-A-DR haplotypic frequencies.
A column (named 'AA') with HLA-A alleles, another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{BDR_freq}{a 3 columns dataframe with HLA-B-DR haplotypic frequencies.
A column (named 'BB') with HLA-B alleles, another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{ABDR_freq}{a 4 columns dataframe with HLA-A-B-DR haplotypic frequencies.
A column (named 'AA') with HLA-A alleles,
another column (named 'BB') with HLA-B alleles,
another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}
}
\value{
a percentual value corresponding to cPRA value
}
\description{
This function computes a cPRA value for one candidate with their HLA sensitization
and given HLA allelic and haplotypic frequencies
}
\details{
This function returns a single percentage for cPRA of a single candidate
considering their HLA immunization and HLA allelic and haplotypic frequencies.
By default HLA frequencies used are those from 37.993 Portuguese bone marrow voluntary donors.
}
\author{
Bruno A Lima
}
