% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpra.R
\name{cpras}
\alias{cpras}
\title{cPRA value for multiple candidates}
\usage{
cpras(sensib = acHLA, A_freq = HLA_A, B_freq = HLA_B,
  DR_freq = HLA_DR, AB_freq = HLA_AB, ADR_freq = HLA_ADR,
  BDR_freq = HLA_BDR, ABDR_freq = HLA_ABDR)
}
\arguments{
\item{sensib}{a 2 columns data frame with patients identifications 'ID'
and respective HLA antibodies 'acs'}

\item{A_freq}{a 2 columns dataframe with HLA-A alelic frequencies.
A column (named 'AA') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{B_freq}{a 2 columns dataframe with HLA-B alelic frequencies.
A column (named 'BB') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{DR_freq}{a 2 columns dataframe with HLA-DR alelic frequencies.
A column (named 'DDR') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{AB_freq}{a 3 columns dataframe with HLA-A-B haplotypic frequencies.
A column (named 'AA') with HLA-A alleles, another (named 'BB') with HLA-B alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{ADR_freq}{a 3 columns dataframe with HLA-A-DR haplotypic frequencies.
A column (named 'AA') with HLA-A alleles, another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{BDR_freq}{a 3 columns dataframe with HLA-B-DR haplotypic frequencies.
A column (named 'BB') with HLA-B alleles, another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{ABDR_freq}{a 4 columns dataframe with HLA-A-B-DR haplotypic frequencies.
A column (named 'AA') with HLA-A alleles,
another column (named 'BB') with HLA-B alleles,
another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}
}
\value{
a data frame with 2 columns, one with patients identification ('ids')
and the other  with percentual values corresponding to respective cPRA values ('cpra')
}
\description{
This function computes cPRA values with \code{cpra1} for multiple candidates with their own HLA sensitization
and given HLA allelic and haplotypic frequencies
}
\details{
This function returns percentages for cPRAs of multiple candidates
considering their respective HLA immunization and HLA allelic and haplotypic frequencies.
By default HLA frequencies used are those from 37.993 Portuguese bone marrow voluntary donors.
}
\author{
Bruno A Lima
}
