% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLA_ABDR}
\alias{HLA_ABDR}
\title{HLA-A-B-DR haplotypic frequencies}
\format{A data frame with 2759 rows and 4 columns:
\itemize{
  \item AA: HLA-A alleles
  \item BB: HLA-B alleles
  \item DDR: HLA-DR alleles
  \item freq: haplotypic relative frequencies
  }}
\usage{
HLA_ABDR
}
\description{
Data frame with HLA loci A-B-DR haplotypic frequencies from Portuguese bone marrow volunteer donors
}
\examples{
\dontrun{
 HLA_ABDR
}
}
\references{
Bruno A. Lima, Helena Alves. HLA-A, -C, -B, and -DRB1 allelic and haplotypic diversity in bone marrow volunteer donors from Northern Portugal. Organs, Tissues & Cells. Volume 16(1), 2013, March: 19-26
 (\href{http://bioestatisticas.wixsite.com/bioestatisticas/publicacoes}{OB})
}
\keyword{datasets}
