% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLA_B}
\alias{HLA_B}
\title{HLA-B allelic frequencies}
\format{A data frame with 34 rows and 2 columns:
\itemize{
  \item BB: HLA-B alleles
  \item freq: allelic relative frequencies
  }}
\usage{
HLA_B
}
\description{
Data frame with HLA locus B allelic frequencies from Portuguese bone marrow volunteer donors
}
\examples{
\dontrun{
 HLA_B
}
}
\references{
Bruno A. Lima, Helena Alves. HLA-A, -C, -B, and -DRB1 allelic and haplotypic diversity in bone marrow volunteer donors from Northern Portugal. Organs, Tissues & Cells. Volume 16(1), 2013, March: 19-26
 (\href{http://bioestatisticas.wixsite.com/bioestatisticas/publicacoes}{OB})
}
\keyword{datasets}
