% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txs.R
\name{txIdents}
\alias{txIdents}
\title{Transplantability percentage for multiple candidates, AB0 **identical**}
\usage{
txIdents(sensib = acHLA, grupos = ABO,
  A_freq = read.csv2("data/hla_A.csv"),
  B_freq = read.csv2("data/hla_B.csv"),
  DR_freq = read.csv2("data/hla_DRB1.csv"),
  AB_freq = read.csv2("data/hla_AB.csv"),
  ADR_freq = read.csv2("data/hla_ADR.csv"),
  BDR_freq = read.csv2("data/hla_BDR.csv"),
  ABDR_freq = read.csv2("data/hla_ABDR.csv"),
  ABO_freq = read.csv2("data/AB0.csv"))
}
\arguments{
\item{sensib}{a 2 columns data frame with patients' identification 'ID'
and HLA antibodies 'acs'}

\item{grupos}{a 2 columns data frame with patients' identification 'ID'
and AB0 blod type 'abo'}

\item{A_freq}{a 2 columns dataframe with HLA-A alelic frequencies.
A column (named 'AA') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{B_freq}{a 2 columns dataframe with HLA-B alelic frequencies.
A column (named 'BB') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{DR_freq}{a 2 columns dataframe with HLA-DR alelic frequencies.
A column (named 'DDR') with HLA-A alleles and a column (named 'freq') with respective frequencies}

\item{AB_freq}{a 3 columns dataframe with HLA-A-B haplotypic frequencies.
A column (named 'AA') with HLA-A alleles, another (named 'BB') with HLA-B alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{ADR_freq}{a 3 columns dataframe with HLA-A-DR haplotypic frequencies.
A column (named 'AA') with HLA-A alleles, another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{BDR_freq}{a 3 columns dataframe with HLA-B-DR haplotypic frequencies.
A column (named 'BB') with HLA-B alleles, another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{ABDR_freq}{a 4 columns dataframe with HLA-A-B-DR haplotypic frequencies.
A column (named 'AA') with HLA-A alleles,
another column (named 'BB') with HLA-B alleles,
another (named 'DDR') with HLA-DR alleles
and a column (named 'freq') with respective haplotypic frequencies}

\item{ABO_freq}{a 2 columns data frame with AB0 blood types frequencies;
A column (named 'grupo') with AB0 blood types,
another column (named 'freq') with respective frequencies.}
}
\value{
a 4 columns data frame with patients' identification ('ID'),
respective cPRA value computed withn \code{cpras} ('cPRA'), compatible transplantability percentage ('TxComp'),
and AB0 blood type ('abo')
}
\description{
This function computes with \code{txIdent1} AB0 identical transplantability percentual scores for multiple cnadidates
given their own HLA sensitization, it's AB0 blood group, and HLA and AB0 genetic frequencies
}
\details{
This function returns percentage transplantability of multiple candidates
considering AB0 compatible, candidates' own HLA immunization and HLA and AB0 genetic frequencies.
By default HLA frequencies used are those from 37.993 Portuguese bone marrow voluntary donors;
and AB0 frequencies are from Portuguese blood donors.
}
\author{
Bruno A Lima
}
